Fix:1.Check plan quota is null;2.Check policy exists when create new policy.

+++ src/main/java/com/asiainfo/bdx/ldp/datafoundry/servicebroker/ocdp/service/OCDPServiceInstanceBindingService.java
@@ -184,8 +184,12 @@ public class OCDPServiceInstanceBindingService implements ServiceInstanceBinding
 
     private void removeServiceInstanceCredentialsItem(ServiceInstance instance, String key){
         Map<String, Object> credentials = instance.getServiceInstanceCredentials();
+		if (! credentials.containsKey(key)){
+			return;
+		}
         credentials.remove(key);
         instance.setCredential(credentials);
+		logger.debug("service instance credentials update to: " + credentials);
 		// delete old service instances
 		repository.delete(instance.getServiceInstanceId());
 		// save new service instance
